// src/auth/enums/auth-type.enum.ts
export enum AuthType {
  Bearer,
  None,
}

export enum Resource {
  Individual = "customers-individual",
  StudentActivity = "customers-studentActivity",
  Company = "customers-company",
  User = "settings-user",
  ExpensesPlace = "settings-expenses-place",
  ExpensesSalaries = "settings-expenses-salaries",
  Product = "kitchen-product",
  Category = "kitchen-category",
  Purchases = "kitchen-purchases",
  Returns = "kitchen-returns",
  Order = "kitchen-orders",
}